use sync state instead of remote heads in SyncIndicator
