+++ src/components/SyncIndicator.tsx
@@ -1,7 +1,5 @@
-import { arraysEqual } from "@/lib/utils"
 import * as A from "@automerge/automerge"
-import { DocHandle } from "@automerge/automerge-repo"
-import { useDocument } from "@automerge/automerge-repo-react-hooks"
+import { DocHandle, PeerId } from "@automerge/automerge-repo"
 import { useState, useEffect } from "react"
 import {
   Circle,
@@ -9,34 +7,42 @@ import {
 } from "lucide-react";
 
 
-const SYNC_SERVER_PEER_ID = "storage-server-sync-automerge-org"
-
 export const SyncIndicator = ({
   handle
 }: {
   handle: DocHandle<unknown>
 }) => {
-  const remoteHeads = useRemoteHeads(handle)
-  const syncServerHeads = remoteHeads[SYNC_SERVER_PEER_ID]?.heads ?? []
-  const currentHeads = handle.docSync() ? A.getHeads(handle.docSync()) : []
-
-  const isSynced = arraysEqual(currentHeads, syncServerHeads)
+  const isSynced = useIsSyncedWithServer(handle)
 
   return isSynced ? <Circle size={"14px"} /> : <CircleDashed size={"14px"} className="" />
 }
 
-function useRemoteHeads(handle) {
-  const [headsMap, setHeadsMap] = useState({})
+const SYNC_SERVER_PEER_ID = "storage-server-sync-automerge-org" as PeerId
+
+function useIsSyncedWithServer (handle: DocHandle<unknown>) : boolean {
+  const [currentHeads, setCurrentHeads] = useState(A.getHeads(handle.docSync()))
+  const [syncedHeads, setSyncedHeads] = useState(handle.getSyncState(SYNC_SERVER_PEER_ID)?.sharedHeads ?? [])
 
   useEffect(() => {
-    handle.on("remote-heads", (data) => {
-      setHeadsMap((headsMap) => ({
-        ...headsMap,
-        [data.remote]: { heads: data.heads, received: data.received }
-      }))
+    handle.on("change", (doc) => {
+      setCurrentHeads(A.getHeads(handle.docSync()))
+    })
+
+    handle.on("sync-state", ({ peerId, syncState }) => {
+      if (peerId === SYNC_SERVER_PEER_ID) {
+        setSyncedHeads(syncState.sharedHeads)
+      }
     })
   }, [handle])
 
 
-  return headsMap
+  return arraysEqual(currentHeads, syncedHeads)
+}
+
+export function arraysEqual(a, b) {
+  if (a.length !== b.length) return false;
+  for (let i = 0; i < a.length; i++) {
+    if (a[i] !== b[i]) return false;
+  }
+  return true;
 }
